Introduced container-native env for make
In CI, it makes sense to run tests natively in pre-provisioned containers, i.e. without using docker-compose to spin up containers ourselves. If we can assume we are already running in an appropriate builder container, we should be able to simply run the test commands (plus whatever is necessary to prepare for running the tests - e.g. setting up databases, etc.)

To accomplish this, I am adding a `container.mk` which will be included in the Makefile, if we are running natively, while maintaining the old behaviour by `refactor -> move` some of the existing make targets to the already existing `docker-compose.mk`.

In this new `container.mk` we don't necessarily need all new targets that the `docker-compose.mk` will have (i.e. the ones we needed for docker-compose)
Some targets which were primarily for manipulating the containers (e.g. build, cache, clean, clean-cache, tmp-export) are not needed. This same functionality will be shifted to the CI config, so that, e.g. database containers are already running and connectivity is provided.